backend/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
xgboost
joblib
psycopg2-binary
sqlalchemy
lightgbm
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
python-multipart
email-validator
